+++ src/js/index.ts
@@ -1,0 +1,43 @@
+export {
+  Breadcrumb,
+  Request,
+  SdkInfo,
+  Event,
+  Exception,
+  Response,
+  Severity,
+  StackFrame,
+  Stacktrace,
+  Status,
+  Thread,
+  User
+} from "@sentry/types";
+
+export {
+  addGlobalEventProcessor,
+  addBreadcrumb,
+  captureException,
+  captureEvent,
+  captureMessage,
+  configureScope,
+  getHubFromCarrier,
+  getCurrentHub,
+  Hub,
+  Scope,
+  setContext,
+  setExtra,
+  setExtras,
+  setTag,
+  setTags,
+  setUser,
+  Span,
+  withScope
+} from "@sentry/core";
+
+export { ReactNativeBackend, ReactNativeOptions } from "./backend";
+export { ReactNativeClient } from "./client";
+export { init, setDist, setRelease } from "./sdk";
+export { SDK_NAME, SDK_VERSION } from "./version";
+
+import * as Integrations from "./integrations";
+export { Integrations };